Instructor/week11-codelab/requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==2.1.0
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
idna==2.8
isort==4.3.4
jieba3k==0.35.1
jsonrpclib-pelix==0.3.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
lxml==4.2.5
Markdown==3.0.1
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.4
numpy==1.15.4
pika==0.12.0
Pillow==5.3.0
pkg-resources==0.0.0
protobuf==3.6.1
pylint==2.2.2
pymongo==3.7.2
python-dateutil==2.7.5
PyYAML==3.13
redis==3.0.1
requests==2.21.0
requests-file==1.4.3
scikit-learn==0.20.2
scipy==1.2.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
tensorboard==1.12.1
tensorflow==1.12.0
termcolor==1.1.0
tinysegmenter==0.3
tldextract==2.2.0
typed-ast==1.1.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
